@@ -12,18 +12,18 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import sys
-
 from google.api_core.gapic_v1 import client_info
 from google.api_core.gapic_v1 import config
+from google.api_core.gapic_v1 import config_async
 from google.api_core.gapic_v1 import method
+from google.api_core.gapic_v1 import method_async
 from google.api_core.gapic_v1 import routing_header
 
-__all__ = ["client_info", "config", "method", "routing_header"]
-
-if sys.version_info >= (3, 6):
-    from google.api_core.gapic_v1 import config_async  # noqa: F401
-    from google.api_core.gapic_v1 import method_async  # noqa: F401
-
-    __all__.append("config_async")
-    __all__.append("method_async")
+__all__ = [
+    "client_info",
+    "config",
+    "config_async",
+    "method",
+    "method_async",
+    "routing_header",
+]
