@@ -33,7 +33,7 @@ class ClientInfo(client_info.ClientInfo):
 
     Args:
         python_version (str): The Python interpreter version, for example,
-            ``'2.7.13'``.
+            ``'3.9.6'``.
         grpc_version (Optional[str]): The gRPC library version.
         api_core_version (str): The google-api-core library version.
         gapic_version (Optional[str]): The sversion of gapic-generated client
