@@ -14,11 +14,7 @@
 
 """Package for interacting with the google.longrunning.operations meta-API."""
 
-import sys
-
+from google.api_core.operations_v1.operations_async_client import OperationsAsyncClient
 from google.api_core.operations_v1.operations_client import OperationsClient
 
-__all__ = ["OperationsClient"]
-if sys.version_info >= (3, 6, 0):
-    from google.api_core.operations_v1.operations_async_client import OperationsAsyncClient  # noqa: F401
-    __all__.append("OperationsAsyncClient")
+__all__ = ["OperationsAsyncClient", "OperationsClient"]
