@@ -15,17 +15,13 @@
 """Helpers for :mod:`protobuf`."""
 
 import collections
+import collections.abc
 import copy
 import inspect
 
 from google.protobuf import field_mask_pb2
 from google.protobuf import message
 from google.protobuf import wrappers_pb2
-
-try:
-    from collections import abc as collections_abc
-except ImportError:  # Python 2.7
-    import collections as collections_abc
 
 
 _SENTINEL = object()
@@ -179,7 +175,7 @@ def get(msg_or_dict, key, default=_SENTINEL):
     # If we get something else, complain.
     if isinstance(msg_or_dict, message.Message):
         answer = getattr(msg_or_dict, key, default)
-    elif isinstance(msg_or_dict, collections_abc.Mapping):
+    elif isinstance(msg_or_dict, collections.abc.Mapping):
         answer = msg_or_dict.get(key, default)
     else:
         raise TypeError(
@@ -204,7 +200,7 @@ def _set_field_on_message(msg, key, value):
     """Set helper for protobuf Messages."""
     # Attempt to set the value on the types of objects we know how to deal
     # with.
-    if isinstance(value, (collections_abc.MutableSequence, tuple)):
+    if isinstance(value, (collections.abc.MutableSequence, tuple)):
         # Clear the existing repeated protobuf message of any elements
         # currently inside it.
         while getattr(msg, key):
@@ -212,13 +208,13 @@ def _set_field_on_message(msg, key, value):
 
         # Write our new elements to the repeated field.
         for item in value:
-            if isinstance(item, collections_abc.Mapping):
+            if isinstance(item, collections.abc.Mapping):
                 getattr(msg, key).add(**item)
             else:
                 # protobuf's RepeatedCompositeContainer doesn't support
                 # append.
                 getattr(msg, key).extend([item])
-    elif isinstance(value, collections_abc.Mapping):
+    elif isinstance(value, collections.abc.Mapping):
         # Assign the dictionary values to the protobuf message.
         for item_key, item_value in value.items():
             set(getattr(msg, key), item_key, item_value)
@@ -241,7 +237,7 @@ def set(msg_or_dict, key, value):
         TypeError: If ``msg_or_dict`` is not a Message or dictionary.
     """
     # Sanity check: Is our target object valid?
-    if not isinstance(msg_or_dict, (collections_abc.MutableMapping, message.Message)):
+    if not isinstance(msg_or_dict, (collections.abc.MutableMapping, message.Message)):
         raise TypeError(
             "set() expected a dict or protobuf message, got {!r}.".format(
                 type(msg_or_dict)
@@ -254,12 +250,12 @@ def set(msg_or_dict, key, value):
     # If a subkey exists, then get that object and call this method
     # recursively against it using the subkey.
     if subkey is not None:
-        if isinstance(msg_or_dict, collections_abc.MutableMapping):
+        if isinstance(msg_or_dict, collections.abc.MutableMapping):
             msg_or_dict.setdefault(basekey, {})
         set(get(msg_or_dict, basekey), subkey, value)
         return
 
-    if isinstance(msg_or_dict, collections_abc.MutableMapping):
+    if isinstance(msg_or_dict, collections.abc.MutableMapping):
         msg_or_dict[key] = value
     else:
         _set_field_on_message(msg_or_dict, key, value)
